Create Plotting_function_in_R_programming
The most used plotting function in R programming is the plot() function. It is a generic function, meaning, it has many methods which are called according to the type of object passed to plot() . In the simplest case, we can pass in a vector and we will get a scatter plot of magnitude vs index.